worked on retrieving and deleting the session from mongo db  and connected the project to mongodb database.
Here i have worked on connecting my project to the mongo db database from where we are going to send and retrieve our data. I made several functions named getSessions, deleteSessions and saveSessions.
I have also removed useraccounts as it was in text and made a json file for the userAccounts data which then i imported to the mongodb database.